混扫任务删除+延时删除板号

+++ bluedragon-distribution-core/src/main/java/com/jd/bluedragon/distribution/jy/service/comboard/impl/JyGroupSortCrossDetailServiceImpl.java
@@ -449,9 +449,20 @@ public class JyGroupSortCrossDetailServiceImpl implements JyGroupSortCrossDetail
         
         // 根据网格删除查询limit外的数据
         for (String groupCode : groupCodeList) {
+            // 查询limit条最新创建的混扫任务的最小ID
+            JyGroupSortCrossDetailEntityQueryDto queryMinIdDto = new JyGroupSortCrossDetailEntityQueryDto();
+            queryMinIdDto.setLimit(limit);
+            queryMinIdDto.setGroupCode(groupCode);
+            Long id = jyGroupSortCrossDetailDao.queryMinIdByGroupCode(queryMinIdDto);
+            if (id == null) {
+                continue;
+            }
+            
+            log.info("开始执行删除网格混扫任务逻辑，网格：{}， id: {}", groupCode, id);
+            // 根据ID删除混扫任务数据
             JyGroupSortCrossDetailEntityQueryDto queryDto = new JyGroupSortCrossDetailEntityQueryDto();
-            queryDto.setLimit(limit);
             queryDto.setGroupCode(groupCode);
+            queryDto.setId(id);
             jyGroupSortCrossDetailDao.deleteCTTGroupDataByGroupCode(queryDto);
         }
     }
